wip: break circular dependencies
