+++ meteor/client/ui/RundownViewKbdShortcuts.tsx
@@ -1,0 +1,29 @@
+export enum RundownViewKbdShortcuts {
+	RUNDOWN_TAKE = 'enter',
+	// RUNDOWN_HOLD = 'h',
+	RUNDOWN_UNDO_HOLD = 'shift+h',
+	RUNDOWN_ACTIVATE = '§',
+	RUNDOWN_ACTIVATE2 = '\\',
+	RUNDOWN_ACTIVATE3 = '|',
+	RUNDOWN_ACTIVATE4 = '½',
+	RUNDOWN_ACTIVATE_REHEARSAL = 'mod+§',
+	RUNDOWN_DEACTIVATE = 'mod+shift+§',
+	RUNDOWN_DEACTIVATE2 = 'mod+shift+½',
+	RUNDOWN_GO_TO_LIVE = 'shift+home',
+	RUNDOWN_REWIND_SEGMENTS = 'mod+home',
+	RUNDOWN_RESET_RUNDOWN = 'shift+escape',
+	RUNDOWN_TOGGLE_SHELF = 'tab',
+	ADLIB_QUEUE_MODIFIER = 'shift',
+	RUNDOWN_NEXT_FORWARD = 'shift+right',
+	RUNDOWN_NEXT_DOWN = 'shift+down',
+	RUNDOWN_NEXT_BACK = 'shift+left',
+	RUNDOWN_NEXT_UP = 'shift+up',
+	RUNDOWN_NEXT_FORWARD2 = 'shift+ArrowRight',
+	RUNDOWN_NEXT_DOWN2 = 'shift+ArrowDown',
+	RUNDOWN_NEXT_BACK2 = 'shift+ArrowLeft',
+	RUNDOWN_NEXT_UP2 = 'shift+ArrowUp',
+	// RUNDOWN_DISABLE_NEXT_ELEMENT = 'g',
+	// RUNDOWN_UNDO_DISABLE_NEXT_ELEMENT = 'shift+g',
+	RUNDOWN_LOG_ERROR = 'shift+backspace',
+	SHOW_CURRENT_SEGMENT_FULL_NONLATCH = '',
+}
