+++ meteor/client/ui/Shelf/AdLibListItem.tsx
@@ -15,7 +15,6 @@ import {
 	LiveSpeakContent,
 	IBlueprintActionTriggerMode,
 } from '@sofie-automation/blueprints-integration'
-import { AdLibPieceUi } from './AdLibPanel'
 import { checkPieceContentStatus } from '../../../lib/mediaObjects'
 import { RundownPlaylist } from '../../../lib/collections/RundownPlaylists'
 import { PubSub } from '../../../lib/api/pubsub'
@@ -28,6 +27,7 @@ import { Studio } from '../../../lib/collections/Studios'
 import { ContextMenuTrigger } from '@jstarpl/react-contextmenu'
 import { contextMenuHoldToDisplayTime, ensureHasTrailingSlash } from '../../lib/lib'
 import { setShelfContextMenuContext, ContextType as MenuContextType } from './ShelfContextMenu'
+import { AdLibPieceUi } from '../../lib/shelf'
 
 export interface IAdLibListItem extends AdLibPieceUi {
 	status: RundownAPI.PieceStatusCode
