+++ meteor/client/ui/Shelf/AdLibRegionPanel.tsx
@@ -7,16 +7,9 @@ import {
 	RundownLayoutAdLibRegionRole,
 } from '../../../lib/collections/RundownLayouts'
 import { RundownLayoutsAPI } from '../../../lib/api/rundownLayouts'
-import {
-	dashboardElementPosition,
-	IDashboardPanelTrackedProps,
-	getUnfinishedPieceInstancesGrouped,
-	getNextPieceInstancesGrouped,
-	isAdLibOnAir,
-	isAdLibNext,
-} from './DashboardPanel'
+import { dashboardElementPosition, IDashboardPanelTrackedProps } from './DashboardPanel'
 import ClassNames from 'classnames'
-import { AdLibPieceUi, IAdLibPanelProps, AdLibFetchAndFilterProps, fetchAndFilter, matchFilter } from './AdLibPanel'
+import { IAdLibPanelProps, AdLibFetchAndFilterProps, fetchAndFilter, matchFilter } from './AdLibPanel'
 import { doUserAction, UserAction } from '../../lib/userAction'
 import { translateWithTracker, Translated } from '../../lib/ReactMeteorData/ReactMeteorData'
 import { MeteorReactComponent } from '../../lib/MeteorReactComponent'
@@ -28,6 +21,13 @@ import { PieceUi } from '../SegmentTimeline/SegmentTimelineContainer'
 import { PieceInstance } from '../../../lib/collections/PieceInstances'
 import { ShowStyleBase } from '../../../lib/collections/ShowStyleBases'
 import { ensureHasTrailingSlash } from '../../lib/lib'
+import {
+	getUnfinishedPieceInstancesGrouped,
+	getNextPieceInstancesGrouped,
+	AdLibPieceUi,
+	isAdLibNext,
+	isAdLibOnAir,
+} from '../../lib/shelf'
 
 interface IState {
 	objId?: string
