+++ meteor/client/ui/Shelf/Inspector/ShelfInspector.tsx
@@ -4,7 +4,6 @@ import { IBlueprintPieceGeneric } from '@sofie-automation/blueprints-integration
 import { IModalAttributes, Modal } from '../../../lib/ui/containers/modals/Modal'
 import renderItem from './ItemRenderers/ItemRendererFactory'
 import { PieceUi } from '../../SegmentTimeline/SegmentTimelineContainer'
-import { AdLibPieceUi } from '../AdLibPanel'
 import { ShowStyleBase } from '../../../../lib/collections/ShowStyleBases'
 import { ContextMenuTrigger } from '@jstarpl/react-contextmenu'
 import { contextMenuHoldToDisplayTime } from '../../../lib/lib'
